test01/requirements.txt
# Requirements for the Prefect Flow project
prefect>=3.4.10
redis>=4.0.0
python-dotenv>=0.19.0
